Create Program35_3.c
This program prints a sequence of characters based on the character entered by the user. If an uppercase letter is entered, it prints all uppercase letters from that character up to 'Z'. If a lowercase letter is entered, it prints all lowercase letters from that character down to 'a'. Non-alphabet characters are printed as they are.