Create Program40_1.c
This program prints a given number of stars using a recursive function named Display. After the user enters a number, the function prints "* " and then calls itself with the value decreased by 1. The recursion continues until iNo becomes 0, at which point the function stops. If the user inputs 5, the program prints: * * * * *.