Create Program40_2.c
This program prints the numbers from 1 to the entered number using recursion. After the user inputs a value, the Display function recursively calls itself with iNo - 1 until it reaches 0. As the recursive calls return, each value of iNo is printed, resulting in the sequence being displayed in ascending order. For an input of 5, the output becomes: 1 2 3 4 5.