Create Program40_3.c
This program prints numbers in descending order using a recursive function. After the user enters a number, the Display function prints the current value of iNo and then calls itself with iNo - 1. This continues until iNo becomes 0, at which point the recursion stops. For an input of 5, the output produced is 5 4 3 2 1.